requirements.txt
fastapi
uvicorn
jinja2
python-multipart
torch
torchvision
numpy
Pillow
open_clip_torch
pymilvus
elasticsearch
google-generativeai
faiss-cpu
sentence-transformers
langdetect
deep-translator
translate
pyvi
underthesea
gdown
requests
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
